# openapi-spec.yaml
openapi: 3.0.3
info:
  title: P4 B2B Marketplace API
  description: API for the P4 GCC/MENA B2B Marketplace
  version: 1.0.0
  contact:
    name: P4 Development Team
    email: dev@p4-marketplace.com
servers:
  - url: https://b2b-marketplace-dcd9azhpefdkdve4.canadacentral-01.azurewebsites.net
    description: Production server
...
paths:
...
  /api/vendors:
    post:
      tags:
        - Catalog
      summary: Create a vendor
      description: Create a new vendor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vendor'
      responses:
        '201':
          description: Vendor created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
    get:
      tags:
        - Catalog
      summary: Get all vendors
      description: Retrieve a list of all vendors with optional filtering and pagination
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            example: "APPROVED"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Vendors retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/vendors/{id}:
    get:
      tags:
        - Catalog
      summary: Get vendor by ID
      description: Retrieve a specific vendor by its ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      responses:
        '200':
          description: Vendor retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '404':
          description: Vendor not found
        '500':
          description: Internal server error
    put:
      tags:
        - Catalog
      summary: Update a vendor
      description: Update an existing vendor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vendor'
      responses:
        '200':
          description: Vendor updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '404':
          description: Vendor not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Catalog
      summary: Delete a vendor
      description: Delete an existing vendor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      responses:
        '200':
          description: Vendor deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '404':
          description: Vendor not found
        '500':
          description: Internal server error
...
  /api/catalog/product-media:
    post:
      tags:
        - Catalog
      summary: Create a product-media association
      description: Create a new product-media association
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductMedia'
      responses:
        '201':
          description: Product-media association created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
    get:
      tags:
        - Catalog
      summary: Get all product-media associations
      description: Retrieve a list of all product-media associations with pagination
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Product-media associations retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/catalog/product-media/{id}:
    get:
      tags:
        - Catalog
      summary: Get product-media association by ID
      description: Retrieve a specific product-media association by its ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      responses:
        '200':
          description: Product-media association retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '404':
          description: Product-media association not found
        '500':
          description: Internal server error
    put:
      tags:
        - Catalog
      summary: Update a product-media association
      description: Update an existing product-media association
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductMedia'
      responses:
        '200':
          description: Product-media association updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '404':
          description: Product-media association not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Catalog
      summary: Delete a product-media association
      description: Delete an existing product-media association
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      responses:
        '200':
          description: Product-media association deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '404':
          description: Product-media association not found
        '500':
          description: Internal server error
  /api/catalog/product-media/product/{productId}:
    get:
      tags:
        - Catalog
      summary: Get media assets for a product
      description: Retrieve all media assets associated with a specific product
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      responses:
        '200':
          description: Media assets retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '404':
          description: Product not found
        '500':
          description: Internal server error
  /api/catalog/products/{id}:
    get:
      tags:
        - Catalog
      summary: Get product by ID
      description: Retrieve a specific product by its ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '404':
          description: Product not found
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        data:
          type: object
          description: Response data payload
          example: {}
        success:
          type: boolean
          description: Whether the request was successful
          example: true
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the response
          example: "2025-10-19T06:30:09.284038836Z"
        metadata:
          type: object
          description: Additional metadata
          example: {"message": "Operation successful"}
        error:
          type: object
          description: Error details if success is false
          properties:
            type:
              type: string
              example: "about:blank"
            title:
              type: string
              example: "Not Found"
            status:
              type: integer
              example: 404
            detail:
              type: string
              example: "Resource not found"
            instance:
              type: string
              example: "/api/resource/123"
    Vendor:
      type: object
      properties:
        id:
          type: string
          example: "01F123456789ABCDEFGH"
        businessName:
          type: string
          example: "Tech Vendor Inc."
        email:
          type: string
          format: email
          example: "contact@techvendor.com"
        phone:
          type: string
          example: "+1234567890"
        address:
          type: string
          example: "123 Tech Street, Tech City, TC 12345"
        taxId:
          type: string
          example: "TV123456789"
        businessLicenseNo:
          type: string
          example: "BL123456789"
        registrationDate:
          type: string
          format: date
          example: "2023-01-01"
        vendorStatus:
          type: string
          enum: [PENDING, APPROVED, REJECTED, SUSPENDED, CLOSED]
          example: "PENDING"
        kycVerified:
          type: boolean
          example: false
        kycVerifiedAt:
          type: string
          format: date
          example: "2023-01-02"
        kycVerifiedBy:
          type: string
          example: "Admin User"
...
    ProductMedia:
      type: object
      properties:
        id:
          type: string
          example: "01F123456789ABCDEFGH"
        productId:
          type: string
          example: "01F987654321ZYXWVU"
        mediaAssetId:
          type: string
          example: "01F555555555555555"
        displayOrder:
          type: integer
          example: 1
        isPrimary:
          type: boolean
          example: true
        altTextOverride:
          type: string
          example: "Primary product image"
...
security:
  - bearerAuth: []